Guessing, timer, and board locking

+++ Wordiple-Client/src/main/java/me/logicologist/wordiple/client/gui/animations/BounceInAnimation.java
@@ -17,7 +17,7 @@ public class BounceInAnimation {
 
     public BounceInAnimation(WritableValue writableValue, int change, int duration) {
         this.writableValue = writableValue;
-        this.originalValue = (int) writableValue.getValue();
+        this.originalValue = (int) (double) writableValue.getValue();
         this.finalValue = originalValue + change;
         this.change = change;
         this.duration = duration;
